Disarm TypeToFocus on mouse clicks
After focusing a window in response to a keystroke, TypeToFocus is disarmed
and waits for the pointer to move into a new window before responding to
any more keyboard activity. Drop into this mode following a mouse/trackpad
click too, which gives more conservative behaviour when a new window is
opened as a result of that click.