Rework the TypeToFocus state machine
When we disarm in response to a mouse click, we should wait for the
mouse to move outside the window under the pointer, not to move outside
the currently focused window. Although the click may focus the window,
we receive the event before it does so.

As we now enter tracking mode following mouse clicks as well as keystrokes,
flipping between the two event taps is likely to be more expensive than a
single callback which returns early for irrelevant events. Consolidate the
taps, simplifying TypeToFocus back towards the original FnToFocus shape.

Resynchronise functions and names which TypeToFocus and FnToFocus share.